Renames model to entity.
