+++ app/src/main/java/com/munstein/gamemanager/entity/Platform.kt
@@ -1,4 +1,4 @@
-package com.munstein.gamemanager.model
+package com.munstein.gamemanager.entity
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
