+++ app/src/main/java/com/munstein/gamemanager/entity/User.kt
@@ -1,3 +1,3 @@
-package com.munstein.gamemanager.model
+package com.munstein.gamemanager.entity
 
 data class User(val id: String, val name: String)
