+++ app/src/main/java/com/munstein/gamemanager/interactor/games/GamesInteractor.kt
@@ -1,6 +1,6 @@
 package com.munstein.gamemanager.interactor.games
 
-import com.munstein.gamemanager.model.Games
+import com.munstein.gamemanager.entity.Games
 import com.munstein.gamemanager.repository.games.IGamesRepository
 
 class GamesInteractor(private val repository: IGamesRepository) : IGamesInteractor {
