+++ app/src/main/java/com/munstein/gamemanager/repository/games/GamesRepository.kt
@@ -1,7 +1,7 @@
 package com.munstein.gamemanager.repository.games
 
 import com.munstein.gamemanager.datasource.FirestoreDataSource
-import com.munstein.gamemanager.model.Games
+import com.munstein.gamemanager.entity.Games
 import kotlinx.coroutines.*
 import kotlinx.coroutines.tasks.await
