+++ app/src/main/java/com/munstein/gamemanager/repository/user/IUserRepository.kt
@@ -1,6 +1,6 @@
 package com.munstein.gamemanager.repository.user
 
-import com.munstein.gamemanager.model.User
+import com.munstein.gamemanager.entity.User
 
 interface IUserRepository {
     fun getUser(): User
